Change: Property definition for better memory use
Previously, each property generator type was redefined on each
call to Modelo.property(). Now these type generator functions are
stored in a submodule scope closure and are simply referenced on
each call to Modelo.property().

Included with this commit are several new assertions to ensure that
the proper scope and context are being maintained with this new
structure.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>