Add: isInstance method to all Modelos
This method can be used to determine if a particular object was
inherited from. The logic behind this is simple for now. It simply
iterates over the objects passed in during a define() or extend()
call and compares each one to the object constructor passed in as
an argument.

Included with this commit are the appropriate tests.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>