Add: Relationship package and tests
This commit contains an extremely rough cut of the relationship
package that is intended for Modelo.

The primary purpose of this commit is to establish an API style
for relationships and to set the expectation that extensions to
Modelo can be created through object extensions rather than direct
modification.

In the long term, this implemenation of relationships will be
replaced with one that is a bit more substantial.

Also included in this commit are the appropriate tests.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>