Change: README to reflect development updates
This commit includes an update to the overal development plan.
Now the order of development is doc, test, code. This pattern
will now be used as this package enters its final phases.

Also included in this commit are updates to the usage examples
in an attempt to give a concise example of usage and directing
readers to the long form docs if they need more.

Also included in this commit are some touchups to the project
description and installation instruction.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>