Change: New keyword to be optional in specs
The specification for Modelo objects now states that instances
can be created with or without the new keyword.

Some of the libraries in this package require the wrapping of
constructor functions within another function to make use of private
variables through scope closures. This, however, would create a
difference between the normal Modelo objects which require the new
keyword and these other objects which do not.

Ultimately, both types of constructors should be available for all
Modelo objects for consistency.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>