Change: Modelo.js to be specification compliant
While no major changes were made to the library itself, there was
a change in the browser export name. Now Modelo is referenced with
a capital letter in the browser.

All tests and documentation have been updated to remain consistent
with the browser export. In addition, this has resolved the JSLint
validation failure caused by using the `new` keyword on a lower
case constructor.

The `new` keyword is optional and has no side-effect on the code
itself. It should be used any time JSLint would, otherwise, note
a failure.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>