Remove: All extra modules
All of the other modules that were contained in this package will
be spun off into their own repositories.

This decision was made to support two development goals. One goal
is to provide developers with a very loose coupling of content so
that they may choose which components of the overall framework to
use in a project.

The second goal is to provide a stress test of the dependency
management module wrapper that has been chosen for this project.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>