Change: Tests to work with new module name
Included with is change is an update to how the tests load the
external dependencies (mocha + expect). Now the 'npm test' runs a
script called 'install_libs' before running the tests. This script
downloads the latest versions of the mocha and expect code, drops
them in the right location for browser testing, and symlinks needed
packages into the node_modules directory in the project root.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>